refactor: homepage components to use client-side rendering
