+++ apps/web/app/home/components/index.tsx
@@ -1,5 +1,15 @@
 import Content from './_Content';
+import { PostReplies } from './_PostReplies';
+import { Timeline } from './_Timeline';
+import { LeftSidebar } from './_LeftSidebar';
+import { RightSidebar } from './_RightSidebar';
+import { MainContent } from './_MainContent';
 
 export const HomePage = {
   Content,
+  PostReplies,
+  Timeline,
+  LeftSidebar,
+  RightSidebar,
+  MainContent
 };
