+++ apps/web/app/home/page.tsx
@@ -1,5 +1,7 @@
 import { getSeoMetadata } from './../../components/HeaderSEO';
+import { Content } from '@social/ui-shared';
 import { HomePage } from './components/index';
+import * as Components from '@/components';
 
 export const metadata = getSeoMetadata({
   title: 'Home | Pubky.app',
@@ -7,5 +9,14 @@ export const metadata = getSeoMetadata({
 });
 
 export default function Index() {
-  return <HomePage.Content />;
+  console.log("here server side")
+  return (
+  <Content.Main>
+      <Components.Header className="hidden md:block" title="Feed" />
+      <Components.RemindBackup />
+      <HomePage.Content />
+      <Components.CreatePost />
+      <Components.FooterMobile title="Feed" />
+  </Content.Main>
+  )
 }
