+++ apps/web/hooks/useIsMobile.ts
@@ -1,0 +1,18 @@
+import { useState, useEffect } from 'react';
+
+export function useIsMobile(breakpoint: number = 1280) {
+  const [isMobile, setIsMobile] = useState(false);
+
+  useEffect(() => {
+    const checkMobile = () => {
+      setIsMobile(window.innerWidth < breakpoint);
+    };
+
+    checkMobile();
+    window.addEventListener('resize', checkMobile);
+
+    return () => window.removeEventListener('resize', checkMobile);
+  }, [breakpoint]);
+
+  return isMobile;
+}
